+++ application/icons/slack-ops-small.png
