+++ application/icons/slack-ops.png
