+++ application/src/.env.example
@@ -1,0 +1,12 @@
+SLACK_BOT_TOKEN=xoxb-xxxxxxxxxxxxx-xxxxxxxxxxxxx-xxxxxxxxxxxxxxxxxxxxxxxx
+SLACK_SIGNING_SECRET=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+DEVOPS_TOKEN=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx=
+APP_VERSION=v0.0.1a
+OPSGENIE_API_KEY=""
+OPSGENIE_SUBSCRIPTION_ID=""
+ORGANIZATION_ID=""
+ORGANIZATION=my-org
+PROJECT=my-proj
+OPS_CHANNEL=""
+OPS_PRIVATE_CHANNEL=""
+PORT=3000
