+++ application/src/approval.py
@@ -44,21 +44,6 @@ def get_intervention_id(build_id, stage_id):
 
     return(json.loads(req.content)["dataProviders"]["ms.vss-build-web.checks-panel-data-provider"][0]["manualValidations"][0]["id"])
 
-###########################################################
-## Simple function to submit an API request to the relevant Release Pipeline,
-## triggering a release to be created with the captured variables
-###########################################################
-def submit_request(request_object, request_type):    
-    headers = AZDO_HEADER
-    devops_pipeline_id = OPS_REQUESTS[request_type]["devops_pipeline_id"]
-    payload = dict(
-        variables = request_object
-    )
-    response = requests.post(PIPELINE_RUN_URL.format(ORGANIZATION, PROJECT, devops_pipeline_id), json = payload, headers = headers )
-    print(response.text)
-    return(response)
-
-
 @app.action("reject")
 ## When the "Reject" button is pressed
 def reject(ack, body, client, logger):
