+++ application/src/requirements.txt
@@ -1,3 +1,4 @@
 slack_bolt
 slack_sdk
 requests
+python-dotenv
