+++ scheduled-tasks/templates/remove_keyvault_policies.yml
@@ -8,6 +8,11 @@ schedules:
       - main
   always: true
 
+variables:
+  - group: Default
+  - name: operation
+    value: 'Scheduled Task'
+    
 steps:
   - template: ../scheduled_tasks.yml
     parameters:
