+++ user-requests/create-new-request.yml
@@ -14,6 +14,9 @@ parameters:
     type: string
     displayName: Type of Pipeline to Create
 
+variables:
+  - group: Default
+
 name: $(Date:yyyyMMdd)-${{ parameters.pipeline_name }}
 
 jobs:
