+++ user-requests/templates/key-vault-policy-add.yml
@@ -1,5 +1,10 @@
 trigger: none
 
+variables:
+  - group: Default
+  - name: operation
+    value: 'User Request'
+    
 steps:
   - template: ../perform_request.yml
     parameters:
